+++ packages/health/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* Updated documentation for Android and Google Fit.
+
+
 ## 1.0.1
 
 * Streamlined DataType units in Flutter.